requirements.txt
numpy==1.20.3
scipy==1.6.3
pyttsx3==2.90
imutils==0.5.4
torch==1.8.1+cu111
opencv_python==4.5.2.52
dlib==19.22.0
Pillow==8.3.0
PyQt5==5.15.4
